Moved components to a sub directory
